Cocoa fixes: * BUG-3945 - Alt-cam bug fix * Use idiomatic cocoa elsewhere in llwindowmaxosx-objc * Back out aura's "fix" * Finish replacing floats with doubles to get mouse deltas

+++ indra/llwindow/llwindowmacosx.h
@@ -124,7 +124,7 @@ public:
 	LLPreeditor* getPreeditor() { return mPreeditor; }
 	
 	void updateMouseDeltas(double* deltas);
-	void getMouseDeltas(float* delta);
+	void getMouseDeltas(double* delta);
 	
 	void handleDragNDrop(std::string url, LLWindowCallbacks::DragNDropAction action);
